+++ src/main/java/com/letsTravel/LetsTravel/domain/type/PrimaryTypeUpdateDTO.java
@@ -7,6 +7,6 @@ import lombok.Getter;
 @AllArgsConstructor
 public class PrimaryTypeUpdateDTO {
 
-	private int placeSeq;
+	private long placeSeq;
 	private String primaryType;
 }
